requirements.txt
altair==6.0.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.2
bs4==0.0.2
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
folium==0.20.0
fonttools==4.61.1
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.15.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
playwright==1.57.0
plotly==6.5.2
prompt_toolkit==3.0.52
protobuf==6.33.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data==0.6.3
streamlit==1.53.1
streamlit-folium==0.26.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
watermark==2.6.0
wcwidth==0.2.14
xyzservices==2025.11.0
zipp==3.23.0
